+++ CHANGELOG.md
@@ -3,6 +3,7 @@ docxBox Changelog
 
 V0.0.3 - Not finished yet
 -------------------------
+* Add unzip and indent XML (uzi)
 * Improve list images (lsi): output file attributes
 
 V0.0.2 - 2020-04-22
@@ -16,11 +17,11 @@ V0.0.1 - 2020-04-20
 * Add replace images of same kind (rpi)
 * Add replace text by text (rpt)
 * Add output plaintext (txt)
-* Add list contents (ls)
 * Add list images (lsi)
 * Add list fields (lsd)
 * Add list fonts (lsf)
 * Add list meta data (lsm)
+* Add list contents (ls)
 * Add zip (zip)
 * Add unzip and unzip media (uz, uzm)
 * Implement application base